+++ programs/folio/src/lib.rs
@@ -45,8 +45,8 @@ pub mod folio {
     #[allow(clippy::too_many_arguments)]
     pub fn init_folio(
         ctx: Context<InitFolio>,
-        folio_fee: u64,
-        minting_fee: u64,
+        folio_fee: u128,
+        minting_fee: u128,
         trade_delay: u64,
         auction_length: u64,
         name: String,
@@ -74,8 +74,8 @@ pub mod folio {
         ctx: Context<UpdateFolio>,
         program_version: Option<Pubkey>,
         program_deployment_slot: Option<u64>,
-        folio_fee: Option<u64>,
-        minting_fee: Option<u64>,
+        folio_fee: Option<u128>,
+        minting_fee: Option<u128>,
         trade_delay: Option<u64>,
         auction_length: Option<u64>,
         fee_recipients_to_add: Vec<FeeRecipient>,
@@ -201,8 +201,8 @@ pub mod folio {
         trade_id: u64,
         sell_limit: Range,
         buy_limit: Range,
-        start_price: u64,
-        end_price: u64,
+        start_price: u128,
+        end_price: u128,
         ttl: u64,
     ) -> Result<()> {
         approve_trade::handler(
@@ -218,10 +218,10 @@ pub mod folio {
 
     pub fn open_trade<'info>(
         ctx: Context<'_, '_, 'info, 'info, OpenTrade<'info>>,
-        sell_limit: u64,
-        buy_limit: u64,
-        start_price: u64,
-        end_price: u64,
+        sell_limit: u128,
+        buy_limit: u128,
+        start_price: u128,
+        end_price: u128,
     ) -> Result<()> {
         open_trade::handler(ctx, sell_limit, buy_limit, start_price, end_price)
     }
